+++ latexcodec/codec.py
@@ -547,7 +547,7 @@ class LatexUnicodeTable:
                  decode=True, encode=True):
         """Register a correspondence between *unicode_text* and *latex_text*.
 
-        :param str unicode_text: A unicode string.
+        :param str unicode_text: A unicode character.
         :param bytes latex_text: Its corresponding LaTeX translation.
         :param str mode: LaTeX mode in which the translation applies
             (``'text'`` or ``'math'``).
@@ -557,6 +557,7 @@ class LatexUnicodeTable:
         :param bool encode: Whether this translation applies to encoding
             (default: ``True``).
         """
+        assert len(unicode_text) == 1
         if package is not None:
             # TODO implement packages
             pass
@@ -625,6 +626,7 @@ class LatexIncrementalEncoder(lexer.LatexIncrementalEncoder):
         else:
             return b'', bytes_
 
+    # TODO refactor this function
     def get_latex_bytes(self, unicode_, final=False):
         if not isinstance(unicode_, string_types):
             raise TypeError(
@@ -632,8 +634,27 @@ class LatexIncrementalEncoder(lexer.LatexIncrementalEncoder):
                 .format(unicode_.__class__.__name__))
         # convert character by character
         for pos, c in enumerate(unicode_):
-            # attempt input encoding first
-            # if this succeeds, then we don't need a latex representation
+            # if ascii, try latex equivalents
+            # (this covers \, #, &, and other special LaTeX characters)
+            if ord(c) < 128:
+                try:
+                    bytes_, tokens = self.table.latex_map[c]
+                except KeyError:
+                    pass
+                else:
+                    # translation succeeded
+                    space, bytes_ = self.get_space_bytes(bytes_)
+                    # update state
+                    if tokens[-1].name == 'control_word':
+                        # we're eating spaces
+                        self.state = 'S'
+                    else:
+                        self.state = 'M'
+                    if space:
+                        yield space
+                    yield bytes_
+                    continue
+            # next, try input encoding
             try:
                 bytes_ = c.encode(self.inputenc, 'strict')
             except UnicodeEncodeError:
@@ -645,8 +666,7 @@ class LatexIncrementalEncoder(lexer.LatexIncrementalEncoder):
                     yield space
                 yield bytes_
                 continue
-            # inputenc failed; let's try the latex equivalents
-            # of common unicode characters
+            # next, try latex equivalents of common unicode characters
             try:
                 bytes_, tokens = self.table.latex_map[c]
             except KeyError: